Don't use POSIX's strtol/strtod conversion.
POSIX's strtod() uses the locale to determine what should
be the decimal separator. That will sometimes be a comma.

Hard to test against, and very confusing. Instead, we now
use a regex for integers and looks_like_number() for floats.

This also changes the coercion interface a bit, requiring only
a single return value.