Fix Enum validation - closes #10
Enum validation finds potential matches to an argument by performing
prefix comparison against the list of valid values. This works fine
as long as the valid values are not prefixes of each other; if they
are, then the shorter strings will never be considered valid input.

This patch fixes that.

Reported by: djerius@cpan.org.