GNU ReadLine - exclude prompt ornaments for leading and trailing whitespace.
The Term::ReadLine::GNU implementation will just slap the prompt between
the ornament-start/ornament-end sequences, but this looks ugly if there
are leading/trailing spaces and the ornament is set to e.g. underline
or standout. This is now brought in line with how the
Term::ReadLine::Perl implementation handles it, by inserting relevant
escape sequences where necessary.