Work around old Pod::Text::Termcap, closes #12
Versions of `Pod::Text::Termcap` before 4.11 had a bug (#124447) that
resulted in using a screen width of 80, regardless of what was specified
in the constructor call or environment. As a result, the high `=over`
indents could cause very strange output or even make Pod::Text::Termcap
spin into an eternal loop.

To make life easier for those with older Perl installations, we test for
the older `Pod::Text::Termcap` version and use `Pod::Text::Overstrike`
instead if necessary. The overstrike variant does not contain the
screen width bug, although it does require a pager that understands
overstrike sequences. The default pager for `Term::CLI` (`less`) does,
so the invonvenience should be minor.