Fix escape sequences during completion (fixes #29)
Unquoted, double quoted, and single quoted partial input
needs to have spaces and quotes handled correctly.

Basically, a "\" in unquoted and double quoted words will
always escape the the next character; in a single quoted
word it has no effect (meaning it is a literal "\").

When adding escape sequences, an unquoted word should have
all "\", double/single quote, and whitespace escaped; for
double quoted words, it's the same, except that whitespace
and single quotes do not need to be escaped; for single
quoted strings, no escaping is done.